server/project/apps/users/serializers/UserSerializer.py
from rest_framework import serializers
from ..models import User

# User serializer
class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'email', 'first_name', 'last_name']
