server/requirements.txt
django
django-cors-headers
djangorestframework
google-auth-oauthlib
google-auth-httplib2
stripe
psycopg2
python-dotenv

